Fixes `LocalNotification` issue that was trying to initialise an `int` with a `double`

+++ lib/src/public/LocalNotification.dart
@@ -13,7 +13,7 @@ class LocalNotification {
   final String body;
 
   /// The delay to the notification in milliseconds.
-  final int delay;
+  final double delay;
 
   LocalNotification({
     required this.type,
